make viewport fix 10inch screen better, create interface for handling touch, torch all mouse hover logic game broken right now
