+++ TowerDefense/GameObjects/Towers/Tower.cs
@@ -12,17 +12,17 @@ namespace TowerDefense
         public Tower(Texture2D tex, Point position) : base(tex, position) { }
 
         public abstract List<Projectile> Attack(List<Enemy> enemylist, List<Projectile> projectilelist, double elapsedTime, Action<int, Point> damageFunc);
-        public override void HandleLeftClick(MouseHandler mouse)
+        public override void HandleLeftClick(InputHandler mouse)
         {
-            switch (mouse.HoveringContext)
-            {
-                case HoveringContext.Tower:
-                    Tower t = mouse.HoveredObject as Tower;
-                    mouse.SelectedObject = t;
-                    mouse.SelectionContext = SelectionContext.TowerSelected;
-                    t.Selected = true;
-                    break;
-            }
+            //switch (mouse.HoveringContext)
+            //{
+            //    case HoveringContext.Tower:
+            //        Tower t = mouse.HoveredObject as Tower;
+            //        mouse.SelectedObject = t;
+            //        mouse.SelectionContext = SelectionContext.TowerSelected;
+            //        t.Selected = true;
+            //        break;
+            //}
         }
 
         public override void ShowStats(SpriteBatch batch)
