+++ TowerDefense/UI/Button.cs
@@ -5,45 +5,46 @@ namespace TowerDefense
 {
     class Button : GameObject
     {
-        //This is sort of a hack to get the button to display the button item's stats given how I have that structured
-        GameObject instanceOfWhatThisButtonCreates;
-        public HoveringContext HoveringContext { get; set; }
-
-        public Button(Point position, Texture2D tex, HoveringContext hoveringContext) : base(tex, position)
+        public Button(Point position, Texture2D tex) : base(tex, position)
         {
-            HoveringContext = hoveringContext;
-            switch (hoveringContext)
+            //HoveringContext = hoveringContext;
+            //switch (hoveringContext)
+            //{
+            //    case HoveringContext.ButtonGenericTower:
+            //        instanceOfWhatThisButtonCreates = new GenericTower(position, tex);
+            //        break;
+            //    case HoveringContext.ButtonCannonTower:
+            //        instanceOfWhatThisButtonCreates = new CannonTower(position, tex);
+            //        break;
+            //    case HoveringContext.ButtonBatteryTower:
+            //        instanceOfWhatThisButtonCreates = new BatteryTower(position, tex);
+            //        break;
+            //    case HoveringContext.ButtonBlastTower:
+            //        instanceOfWhatThisButtonCreates = new BlastTower(position, tex);
+            //        break;
+            //    case HoveringContext.ButtonWall:
+            //        instanceOfWhatThisButtonCreates = new Node(Point.Zero, Point.Zero, tex) { wall = true };
+            //        break;
+
+            //Big hack, fix this
+            if (tex == ResourceManager.Portal)
             {
-                case HoveringContext.ButtonGenericTower:
-                    instanceOfWhatThisButtonCreates = new GenericTower(position, tex);
-                    break;
-                case HoveringContext.ButtonCannonTower:
-                    instanceOfWhatThisButtonCreates = new CannonTower(position, tex);
-                    break;
-                case HoveringContext.ButtonBatteryTower:
-                    instanceOfWhatThisButtonCreates = new BatteryTower(position, tex);
-                    break;
-                case HoveringContext.ButtonBlastTower:
-                    instanceOfWhatThisButtonCreates = new BlastTower(position, tex);
-                    break;
-                case HoveringContext.ButtonWall:
-                    instanceOfWhatThisButtonCreates = new Node(Point.Zero, Point.Zero, tex) { wall = true };
-                    break;
-                case HoveringContext.ButtonPortal:
-                    instanceOfWhatThisButtonCreates = new Node(Point.Zero, Point.Zero, tex) { portal = true };
-                    SpriteHeight = 32;
-                    SpriteWidth = 32;
-                    currentFrame = 0;
-                    frameCount = 4;
-                    break;
-                case HoveringContext.ButtonCheese:
-                    instanceOfWhatThisButtonCreates = new Node(Point.Zero, Point.Zero, tex) { cheese = true };
-                    break;
-                case HoveringContext.ButtonUpgrade:
-                    break;
-                default:
-                    break;
+                SpriteHeight = 32;
+                SpriteWidth = 32;
+                currentFrame = 0;
+                frameCount = 4;
             }
+            //        instanceOfWhatThisButtonCreates = new Node(Point.Zero, Point.Zero, tex) { portal = true };
+
+            //        break;
+            //    case HoveringContext.ButtonCheese:
+            //        instanceOfWhatThisButtonCreates = new Node(Point.Zero, Point.Zero, tex) { cheese = true };
+            //        break;
+            //    case HoveringContext.ButtonUpgrade:
+            //        break;
+            //    default:
+            //        break;
+            //}
         }
 
         public override void Draw(SpriteBatch batch)
@@ -53,7 +54,7 @@ namespace TowerDefense
 
         public override void ShowStats(SpriteBatch batch)
         {
-            instanceOfWhatThisButtonCreates?.ShowStats(batch);
+            //instanceOfWhatThisButtonCreates?.ShowStats(batch);
         }
 
         public override Rectangle BoundingBox()
@@ -61,57 +62,57 @@ namespace TowerDefense
             return new Rectangle(Position.X - SpriteWidth / 2, Position.Y - SpriteHeight / 2, SpriteWidth, SpriteHeight);
         }
 
-        public override void HandleLeftClick(MouseHandler mouse)
+        public override void HandleLeftClick(InputHandler mouse)
         {
-            switch (mouse.HoveringContext)
-            {
-                case HoveringContext.ButtonGenericTower:
-                    mouse.UpdateTex(mouse.HoveredObject.Tex);
-                    mouse.SelectedObject = new GenericTower(mouse.Position, mouse.Tex);
-                    mouse.SelectionContext = SelectionContext.PlacingTower;
-                    break;
-                case HoveringContext.ButtonCannonTower:
-                    mouse.UpdateTex(mouse.HoveredObject.Tex);
-                    mouse.SelectedObject = new CannonTower(mouse.Position, mouse.Tex);
-                    mouse.SelectionContext = SelectionContext.PlacingTower;
-                    break;
-                case HoveringContext.ButtonBatteryTower:
-                    mouse.UpdateTex(mouse.HoveredObject.Tex);
-                    mouse.SelectedObject = new BatteryTower(mouse.Position, mouse.Tex);
-                    mouse.SelectionContext = SelectionContext.PlacingTower;
-                    break;
-                case HoveringContext.ButtonBlastTower:
-                    mouse.UpdateTex(mouse.HoveredObject.Tex);
-                    mouse.SelectedObject = new BlastTower(mouse.Position, mouse.Tex);
-                    mouse.SelectionContext = SelectionContext.PlacingTower;
-                    break;
-                case HoveringContext.ButtonWall:
-                    mouse.UpdateTex(mouse.HoveredObject.Tex);
-                    mouse.SelectionContext = SelectionContext.PlacingWall;
-                    break;
-                case HoveringContext.ButtonPortal:
-                    mouse.UpdateTex(mouse.HoveredObject.Tex, 32, 32, 4);
-                    mouse.SelectionContext = SelectionContext.PlacingPortalEntrance;
-                    break;
-                case HoveringContext.ButtonCheese:
-                    mouse.UpdateTex(mouse.HoveredObject.Tex);
-                    mouse.SelectionContext = SelectionContext.PlacingCheese;
-                    break;
-                case HoveringContext.ButtonUpgrade:
-                    Tower t = mouse.SelectedObject as Tower;
-                    if (GameStats.Gold >= t.Cost)
-                    {
-                        GameStats.Gold = GameStats.Gold - t.Cost;
-                        t.upgrade();
-                    }
-                    else
-                    {
-                        MessageLog.NotEnoughGold();
-                    }
-                    break;
-                default:
-                    break;
-            }
+            //switch (mouse.HoveringContext)
+            //{
+            //    case HoveringContext.ButtonGenericTower:
+            //        mouse.UpdateTex(mouse.HoveredObject.Tex);
+            //        mouse.SelectedObject = new GenericTower(mouse.Position, mouse.Tex);
+            //        mouse.SelectionContext = SelectionContext.PlacingTower;
+            //        break;
+            //    case HoveringContext.ButtonCannonTower:
+            //        mouse.UpdateTex(mouse.HoveredObject.Tex);
+            //        mouse.SelectedObject = new CannonTower(mouse.Position, mouse.Tex);
+            //        mouse.SelectionContext = SelectionContext.PlacingTower;
+            //        break;
+            //    case HoveringContext.ButtonBatteryTower:
+            //        mouse.UpdateTex(mouse.HoveredObject.Tex);
+            //        mouse.SelectedObject = new BatteryTower(mouse.Position, mouse.Tex);
+            //        mouse.SelectionContext = SelectionContext.PlacingTower;
+            //        break;
+            //    case HoveringContext.ButtonBlastTower:
+            //        mouse.UpdateTex(mouse.HoveredObject.Tex);
+            //        mouse.SelectedObject = new BlastTower(mouse.Position, mouse.Tex);
+            //        mouse.SelectionContext = SelectionContext.PlacingTower;
+            //        break;
+            //    case HoveringContext.ButtonWall:
+            //        mouse.UpdateTex(mouse.HoveredObject.Tex);
+            //        mouse.SelectionContext = SelectionContext.PlacingWall;
+            //        break;
+            //    case HoveringContext.ButtonPortal:
+            //        mouse.UpdateTex(mouse.HoveredObject.Tex, 32, 32, 4);
+            //        mouse.SelectionContext = SelectionContext.PlacingPortalEntrance;
+            //        break;
+            //    case HoveringContext.ButtonCheese:
+            //        mouse.UpdateTex(mouse.HoveredObject.Tex);
+            //        mouse.SelectionContext = SelectionContext.PlacingCheese;
+            //        break;
+            //    case HoveringContext.ButtonUpgrade:
+            //        Tower t = mouse.SelectedObject as Tower;
+            //        if (GameStats.Gold >= t.Cost)
+            //        {
+            //            GameStats.Gold = GameStats.Gold - t.Cost;
+            //            t.upgrade();
+            //        }
+            //        else
+            //        {
+            //            MessageLog.NotEnoughGold();
+            //        }
+            //        break;
+            //    default:
+            //        break;
+            //}
         }
     }
 }
